requirements.txt
langchain
openai
python-dotenv
pyyaml
aiogram
polib
solana